services/requirements.txt
fastapi
uvicorn
pydantic-settings
requests
python-dotenv
pydantic[email]
pandas
qdrant-client
sentence-transformers
python-multipart
faker